global: restructure book model
* book model restructure for integration with the rest
  of the migration flavours
