+++ cds_dojson/marc21/models/books/book.py
@@ -16,21 +16,97 @@
 # You should have received a copy of the GNU General Public License
 # along with Invenio; if not, write to the Free Software Foundation, Inc.,
 # 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
-"""Video model."""
+"""Book model"""
 
 from cds_dojson.overdo import OverdoJSONSchema
 from cds_dojson.marc21.fields.base import model as cds_base
+from dojson._compat import iteritems
+from dojson.errors import IgnoreKey, MissingRule
+from dojson.utils import GroupableOrderedDict
 
 
-class CDSBook(OverdoJSONSchema):
+class CDSOverdoBase(OverdoJSONSchema):
+
+    def do(self, blob, ignore_missing=True, exception_handlers=None):
+        """Translate blob values and instantiate new model instance.
+
+        Raises ``MissingRule`` when no rule matched and ``ignore_missing``
+        is ``False``.
+
+        :param blob: ``dict``-like object on which the matching rules are
+                     going to be applied.
+        :param ignore_missing: Set to ``False`` if you prefer to raise
+                               an exception ``MissingRule`` for the first
+                               key that it is not matching any rule.
+        :param exception_handlers: Give custom exception handlers to take care
+                                   of non-standard codes that are installation
+                                   specific.
+
+        .. versionchanged:: 1.0.0
+
+           ``ignore_missing`` allows to specify if the function should raise
+           an exception.
+
+        .. versionchanged:: 1.1.0
+
+           ``exception_handlers`` allows to set custom handlers for
+           non-standard MARC codes.
+        """
+        handlers = {IgnoreKey: None}
+        handlers.update(exception_handlers or {})
+
+        def clean_missing(exc, output, key, value):
+            order = output.get('__order__')
+            if order:
+                order.remove(key)
+
+        if ignore_missing:
+            handlers.setdefault(MissingRule, clean_missing)
+
+        output = {}
+
+        if self.index is None:
+            self.build()
+        if isinstance(blob, GroupableOrderedDict):
+            items = blob.iteritems(repeated=True, with_order=False)
+        else:
+            items = iteritems(blob)
+        for key, value in items:
+            try:
+                result = self.index.query(key)
+                if not result:
+                    raise MissingRule(key)
+
+                name, creator = result
+                data = creator(output, key, value)
+                if getattr(creator, '__extend__', False):
+                    existing = output.get(name, [])
+                    existing.extend(data)
+                    output[name] = existing
+                else:
+                    output[name] = data
+            except Exception as exc:
+                if exc.__class__ in handlers:
+                    handler = handlers[exc.__class__]
+                    if handler is not None:
+                        handler(exc, output, key, value)
+                else:
+                    raise
+        return output
+
+
+class CDSBook(CDSOverdoBase):
     """Translation Index for CDS Videos."""
 
-    __query__ = ''
+    __query__ = '980__:BOOK -980__:PROCEEDINGS -980__:DELETED'
 
     __schema__ = 'records/books/book/book-v.0.0.1.json'
 
     __ignore_keys__ = {
-
+        '940',
+        '964',
+        '981',
+        '963',
     }
 
     def do(self, blob, ignore_missing=True, exception_handlers=None):
