+++ cds_dojson/overdo.py
@@ -19,10 +19,7 @@
 """Base classes for CDS DoJSON."""
 
 import pkg_resources
-from dojson._compat import iteritems
-from dojson.errors import IgnoreKey, MissingRule
 from dojson.overdo import Overdo as DoJSONOverdo
-from dojson.utils import GroupableOrderedDict
 
 from .matcher import matcher
 from .utils import not_accessed_keys
@@ -89,73 +86,6 @@ class Overdo(DoJSONOverdo):
             self.rules[:] = [rule for rule in self.rules if not override(rule)]
 
         return super(Overdo, self).over(name, *source_tags)
-
-    def do(self, blob, ignore_missing=True, exception_handlers=None):
-        """Translate blob values and instantiate new model instance.
-
-        Raises ``MissingRule`` when no rule matched and ``ignore_missing``
-        is ``False``.
-
-        :param blob: ``dict``-like object on which the matching rules are
-                     going to be applied.
-        :param ignore_missing: Set to ``False`` if you prefer to raise
-                               an exception ``MissingRule`` for the first
-                               key that it is not matching any rule.
-        :param exception_handlers: Give custom exception handlers to take care
-                                   of non-standard codes that are installation
-                                   specific.
-
-        .. versionchanged:: 1.0.0
-
-           ``ignore_missing`` allows to specify if the function should raise
-           an exception.
-
-        .. versionchanged:: 1.1.0
-
-           ``exception_handlers`` allows to set custom handlers for
-           non-standard MARC codes.
-        """
-        handlers = {IgnoreKey: None}
-        handlers.update(exception_handlers or {})
-
-        def clean_missing(exc, output, key, value):
-            order = output.get('__order__')
-            if order:
-                order.remove(key)
-
-        if ignore_missing:
-            handlers.setdefault(MissingRule, clean_missing)
-
-        output = {}
-
-        if self.index is None:
-            self.build()
-        if isinstance(blob, GroupableOrderedDict):
-            items = blob.iteritems(repeated=True, with_order=False)
-        else:
-            items = iteritems(blob)
-        for key, value in items:
-            try:
-                result = self.index.query(key)
-                if not result:
-                    raise MissingRule(key)
-
-                name, creator = result
-                data = creator(output, key, value)
-                if getattr(creator, '__extend__', False):
-                    existing = output.get(name, [])
-                    existing.extend(data)
-                    output[name] = existing
-                else:
-                    output[name] = data
-            except Exception as exc:
-                if exc.__class__ in handlers:
-                    handler = handlers[exc.__class__]
-                    if handler is not None:
-                        handler(exc, output, key, value)
-                else:
-                    raise
-        return output
 
     def missing(self, blob, **kwargs):
         """Return keys with missing rules."""
